backend/requirements.txt
Django==3.2.16 
djangorestframework
djangorestframework-simplejwt
django-filter==22.1
django-colorfield==0.10.1
djoser==2.2.0 
drf-extra-fields==3.6.1
Pillow==9.3.0 
PyJWT==2.1.0 
requests==2.26.0
PyYAML==6.0 
python-dotenv==1.0.0 
gunicorn==20.1.0
psycopg2-binary==2.9.3